About Myself clean-up
