--- src/pages/admin/_username/AdminUsernamePage.scss
@@ -1,31 +1,0 @@
-@import 'src/variables';
-
-.admin-username-page {
-  &__date {
-    color: $dark_purple;
-  }
-}
-
-.username {
-  &__avatar {
-    img {
-      border: 10px solid $pink;
-      border-radius: 50%;
-      height: 200px;
-      width: auto;
-    }
-  }
-
-  &__hero {
-    background: $primary;
-    color: white;
-    padding: 20px;
-  }
-
-  &__profile-info {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-  }
-}
